chore: 기존 exception 코드 삭제 및 패키지 이동
