+++ src/main/java/com/ani/taku_backend/common/advice/ApiResponseStatusAdvice.java
@@ -1,6 +1,6 @@
-package com.ani.taku_backend.global.advice;
+package com.ani.taku_backend.common.advice;
 
-import com.ani.taku_backend.global.response.ApiResponse;
+import com.ani.taku_backend.common.response.ApiResponse;
 import org.springframework.core.MethodParameter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
 
 /*
-해당 클래스의 역할은  ApiResponse를 사용하는 응답의 HttpStatus를 의도한대로 바꾸는 것입니다.
+해당 클래스의 역할은 ApiResponse를 사용하는 응답의 HttpStatus를 의도한대로 바꾸는 것입니다.
 */
 
 @RestControllerAdvice
