--- src/main/java/com/ani/taku_backend/common/exception/AlreadyExistsException.java
@@ -1,9 +1,0 @@
-package com.ani.taku_backend.common.exception;
-
-public class AlreadyExistsException extends RuntimeException {
-    private String message;
-
-    public AlreadyExistsException(String message) {
-        super("해당 "+ message + " 은 이미 존재합니다.");
-    }
-}
