+++ src/main/java/com/ani/taku_backend/common/exception/CustomException.java
@@ -1,4 +1,4 @@
-package com.ani.taku_backend.global.exception;
+package com.ani.taku_backend.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
