+++ src/main/java/com/ani/taku_backend/common/exception/GlobalExceptionHandler.java
@@ -1,66 +1,35 @@
 package com.ani.taku_backend.common.exception;
 
-import com.ani.taku_backend.common.ApiConstants;
-import com.ani.taku_backend.common.model.MainResponse;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
+import com.ani.taku_backend.common.response.ApiResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.HttpRequestMethodNotSupportedException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+import org.springframework.web.servlet.NoHandlerFoundException;
 
-@ControllerAdvice
+@Slf4j
+@RestControllerAdvice
 public class GlobalExceptionHandler {
-
-    // SQL 관련
-    @ExceptionHandler(DataIntegrityViolationException.class)
-    public ResponseEntity<MainResponse<Void>> handleDataIntegrityViolation(DataIntegrityViolationException ex) {
-        Throwable rootCause = ex.getRootCause();
-        String errorMessage = (rootCause != null ? rootCause.getMessage() : ex.getMessage());
-        String parseMessage = parseColumnNameAndMessage(errorMessage);
-
-        return ResponseEntity.badRequest().body(
-                new MainResponse<>(
-                        ApiConstants.Status.ERROR,
-                        parseMessage != null ? parseMessage : ApiConstants.Message.BAD_REQUEST
-                )
-        );
+    // 존재하지 않는 요청에 대한 예외
+    @ExceptionHandler(value = {NoHandlerFoundException.class, HttpRequestMethodNotSupportedException.class})
+    public ApiResponse<?> handleNoPageFoundException(Exception e) {
+        log.error("GlobalExceptionHandler catch NoHandlerFoundException : {}", e.getMessage());
+        return ApiResponse.fail(ErrorCode.NOT_FOUND_END_POINT);
     }
 
-    // 공통 Exception
-    @ExceptionHandler(Exception.class)
-    @ResponseBody
-    public ResponseEntity<MainResponse<Void>> handleException(Exception ex) {
-        return ResponseEntity.badRequest().body(
-                new MainResponse<>(
-                        ApiConstants.Status.ERROR,
-                        ex.getMessage() != null ? ex.getMessage() : ApiConstants.Message.BAD_REQUEST
-                )
-        );
+
+    // 커스텀 예외
+    @ExceptionHandler(value = {CustomException.class})
+    public ApiResponse<?> handleCustomException(CustomException e) {
+        log.error("handleCustomException() in GlobalExceptionHandler throw CustomException : {}", e.getMessage());
+        return ApiResponse.fail(e.getErrorCode());
     }
 
-    public String parseColumnNameAndMessage(String errorMessage) {
-        String columnName = "undefined";
-
-        // H2 메시지에서 컬럼명 추출(길이초과)
-        if (errorMessage.contains("Value too long for column")) {
-            int startIndex = errorMessage.indexOf("\"") + 1;
-            int endIndex = errorMessage.indexOf(" ", startIndex);
-            if (startIndex > 0 && endIndex > startIndex) {
-                columnName = errorMessage.substring(startIndex, endIndex).trim();
-            }
-            errorMessage = "입력된 데이터가 허용된 길이를 초과했습니다. (컬럼명: " + columnName.toLowerCase() + ")";
-        }
-
-        // MySQL 메시지에서 컬럼명 추출(길이초과)
-        else if (errorMessage.contains("Data too long for column")) {
-            int startIndex = errorMessage.indexOf("'") + 1;
-            int endIndex = errorMessage.indexOf("'", startIndex);
-            if (startIndex > 0 && endIndex > startIndex) {
-                columnName = errorMessage.substring(startIndex, endIndex).trim();
-            }
-            errorMessage = "입력된 데이터가 허용된 길이를 초과했습니다. (컬럼명: " + columnName.toLowerCase() + ")";
-        }
-
-        return errorMessage;
+    // 기본 예외
+    @ExceptionHandler(value = {Exception.class})
+    public ApiResponse<?> handleException(Exception e) {
+        log.error("handleException() in GlobalExceptionHandler throw Exception : {}", e.getMessage());
+        e.printStackTrace();
+        return ApiResponse.fail(ErrorCode.INTERNAL_SERVER_ERROR);
     }
 }
